h264nal: clear the remaining clang-tidy findings
"make tidy" reported 4 warnings, and had done for a while. None was a
live bug, which is why they sat: each was checked and each is
unreachable or harmless today. They are fixed anyway, for two reasons.
Three of the four are safe only because of an invariant established
somewhere else in the function, so the code is one edit away from being
wrong. And a standing set of 4 known-benign findings is where a fifth,
real one goes unnoticed.

bugprone-implicit-widening-of-multiplication-result
h264_rtp_stapa_parser.cc:57

    if (nalu_size * 8 > bit_buffer->RemainingBitCount()) {

  computes the product in uint32_t and widens it to uint64_t for the
  comparison. It cannot wrap, because nalu_size is filled by
  ReadBits(16, ...) four lines above, so it caps at 65535 and the product
  at 524280. But this is the bounds check on a length taken off the
  network, and what keeps it correct is that 16, not anything local: read
  32 bits there instead and this check silently stops bounding anything,
  failing open. RFC 6184 has aggregation formats this parser does not
  implement yet, MTAP16 and MTAP24, and adding one means writing exactly
  that code here. Cast to uint64_t before the multiply.

bugprone-misplaced-widening-cast
h264_sps_extension_parser.cc:93 and :101

    static_cast<size_t>(sps_extension->bit_depth_aux_minus8 + 9)

  adds in uint32_t and casts the result, so the cast reads as though the
  arithmetic happens in size_t when it does not. The sum is at most 13,
  since bit_depth_aux_minus8 is rejected outside 0 to 4 twenty lines
  earlier, but this is a width fed to ReadBits, which is the pattern
  TODO.md section 2 is about and the one that produced the SVC crash.
  Cast before the addition.

  The two lines computed the same expression twice from a value that
  cannot change between them. Section 7.3.2.1.2 gives both
  alpha_opaque_value and alpha_transparent_value the same u(v) width, so
  compute alpha_value_bits once and use it for both.

bugprone-narrowing-conversions
h264_vui_parameters_parser.cc:334

    float framerate = (float)time_scale / (2.0 * (float)num_units_in_tick);

  The obvious fix is to write 2.0f, and it is the wrong one: that
  silences the check by making the division float, which removes the
  precision rather than acknowledging the narrowing. The 2.0 being a
  double is what makes the division double, and 30000/1001 is the sort of
  ratio that wants it. Make the narrowing explicit instead and leave the
  arithmetic in double. The two (float) casts on the operands go with it:
  both promote straight back to double, and they were the only C-style
  casts of their kind in the tree, against 43 static_casts.

  This one is not a no-op. Over 20 million random operand pairs, 33.6%
  produce a different float, and the new form is the more accurate: the
  old code rounded both operands to float before dividing, which is a
  double rounding.

      exact : 0.175564373331
      before: 0.175564363599  err 9.73e-09
      after : 0.175564378500  err 5.17e-09

  No realistic input is affected. 60000/1001, 50/1 and every plausible
  H.264 timing pair are bit-identical; divergence needs operands past
  float's 2^24 exact integer range.

"make tidy" now reports 0 warnings over the whole tree. gcc and clang
both build with no errors and no warnings, and 23/23 tests pass.

Two gaps these fixes expose but do not close, both noted for later:
h264_rtp_stapa_parser.cc has a single unit test, over a well formed
packet, so the reject path hardened here has no coverage at all; and
num_units_in_tick is read as a raw u(32) with no validation, although
section E.2.1 requires it to be greater than 0, so a zero reaches the
division above and yields an infinity.

+++ src/h264_rtp_stapa_parser.cc
@@ -53,8 +53,12 @@ H264RtpStapAParser::ParseRtpStapA(
     if (!bit_buffer->ReadBits(16, nalu_size)) {
       return nullptr;
     }
-    // Validate nalu_size fits in remaining buffer
-    if (nalu_size * 8 > bit_buffer->RemainingBitCount()) {
+    // Validate nalu_size fits in remaining buffer. Widen before the
+    // multiply, not after: nalu_size is 16 bits today so the product
+    // cannot wrap, but that is a property of the ReadBits above rather
+    // than of this check.
+    if (static_cast<uint64_t>(nalu_size) * 8 >
+        bit_buffer->RemainingBitCount()) {
       return nullptr;
     }
     rtp_stapa->nal_unit_sizes.push_back(nalu_size);
